Move SchedulerXComArg to serialization
